refactor: change packetSNR type from int8_t to float and update print function to include SNR

+++ LoraReceiver.h
@@ -54,7 +54,7 @@
 extern volatile bool is_packet_received;
 extern char receivedMessage[];
 extern int16_t packetRSSI;
-extern int8_t packetSNR;
+extern float packetSNR;
 
 /*****************************************************************/
 /* STRUCTURES                                                    */
